new route setup and individual component rendered based on their route, not only in the dashboard
