+++ frontend/src/components/DashLayout.jsx
@@ -1,9 +1,9 @@
+import React, { useState, useEffect } from "react";
 import { Outlet } from "react-router-dom";
 import Sidebar from "@/components/Sidebar";
 import Header from "@/components/Header";
 import { useQuery } from "@apollo/client";
 import { GET_ME } from "../utils/queries";
-import { useState, useEffect } from "react";
 import { useToast } from "@/components/ui/use-toast";
 import { welcomeMessage } from "@/utils/helpers";
 import Skeleton from "@/components/Skeleton";
@@ -11,15 +11,68 @@ import Skeleton from "@/components/Skeleton";
 const DashLayout = () => {
   const { toast } = useToast();
   const [user, setUser] = useState({});
+  const [chartData, setChartData] = useState([]);
+  const [cardValues, setCardValues] = useState({
+    totalIncome: 0,
+    totalExpenses: 0,
+    balance: 0,
+  });
 
   const { loading, data } = useQuery(GET_ME, {
     pollInterval: 1000,
     nextFetchPolicy: "cache-first",
   });
 
+  const calculateTotals = (expenses) => {
+    let totalIncome = 0;
+    let totalExpense = 0;
+    let totalInvestment = 0;
+
+    expenses.forEach((expense) => {
+      if (expense.category === "income") {
+        totalIncome += expense.amount;
+      } else if (expense.category === "expense") {
+        totalExpense += expense.amount;
+      } else if (expense.category === "investment") {
+        totalInvestment += expense.amount;
+      }
+    });
+
+    const totalExpenses = totalExpense + totalInvestment;
+    const balance = totalIncome - totalExpenses;
+
+    return {
+      totalIncome,
+      totalExpenses,
+      balance,
+      pieData: [
+        {
+          category: "Income",
+          amount: totalIncome,
+          fill: "hsl(var(--chart-2))",
+        },
+        {
+          category: "Expense",
+          amount: totalExpense,
+          fill: "hsl(var(--chart-1))",
+        },
+        {
+          category: "Investment",
+          amount: totalInvestment,
+          fill: "hsl(var(--chart-4))",
+        },
+      ],
+    };
+  };
+
   useEffect(() => {
     if (data && data.me) {
       setUser(data.me);
+      const { totalIncome, totalExpenses, balance, pieData } = calculateTotals(
+        data.me.expenses
+      );
+      setChartData(pieData);
+      setCardValues({ totalIncome, totalExpenses, balance });
 
       if (!welcomeMessage()) {
         toast({
@@ -40,7 +93,7 @@ const DashLayout = () => {
       <div className="flex flex-col sm:gap-4 sm:py-4 sm:pl-14">
         <Header user={user} />
         <main className="p-4 sm:px-6 sm:py-0 md:gap-8 container">
-          <Outlet />
+          <Outlet context={{ user, chartData, cardValues }} />
         </main>
       </div>
     </div>
