+++ frontend/src/components/PieChart.jsx
@@ -1,6 +1,7 @@
 import * as React from "react";
 import { Label, Pie, PieChart } from "recharts";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { useOutletContext } from "react-router-dom";
 import {
   ChartContainer,
   ChartTooltip,
@@ -11,19 +12,24 @@ import {
 const chartConfig = {};
 
 export default function Chart({ data }) {
+  const { chartData } = useOutletContext(); // Access chartData from context
+
+  // Safely calculating totalExpenses, providing 0 as a fallback if chartData is undefined or empty
   const totalExpenses = React.useMemo(() => {
-    return data.reduce((acc, curr) => acc + curr.amount, 0);
-  }, []);
+    return chartData && chartData.length > 0
+      ? chartData.reduce((acc, curr) => acc + curr.amount, 0)
+      : 0;
+  }, [chartData]);
 
   return (
     <Card className="flex flex-col">
       <CardHeader className="items-center pb-0">
         <CardTitle>Pie Chart Overview</CardTitle>
       </CardHeader>
-      <CardContent className="flex-1 pb-0">
+      <CardContent className="flex-1 pb-0  ">
         <ChartContainer
           config={chartConfig}
-          className="mx-auto aspect-square max-h-[400px]"
+          className="mx-auto aspect-square max-h-[250px]"
         >
           <PieChart>
             <ChartTooltip
@@ -31,10 +37,10 @@ export default function Chart({ data }) {
               content={<ChartTooltipContent hideLabel />}
             />
             <Pie
-              data={data}
+              data={chartData}
               dataKey="amount"
               nameKey="category"
-              innerRadius={70}
+              innerRadius={60}
               strokeWidth={5}
             >
               <Label
